refactor: extract render and puml_encoder into shared/ package
- Move render.py and puml_encoder.py into shared/ sub-package
- Fix activity_classes → activity import references
- Update imports in app.py and tests accordingly
