+++ src/plantuml_gui/shared/puml_encoder.py
